requirements.txt
django
python-decouple
djangorestframework
gunicorn
psycopg2-binary
Pillow
django-modeltranslation
drf-yasg
django-cors-headers
django-filter
python-dateutil
instabot==0.117.0
sentry-sdk[django]
django-channels 
channels-redis